requirements.txt
Django==4.1.7
python-dotenv
requests
psycopg2
djangorestframework
coreapi
drf_yasg